requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
jinja2==3.1.2
python-dotenv==1.0.0
airtable-python-wrapper==0.15.3
python-multipart==0.0.6

